refactor(desktop): keep oversized files at their base line counts; drop dead onboarding analytics hooks
